# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Journey API",
    "description": "The public API documentation for the Journey web application.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local host server for development"
    }
  ],
  "paths": {
    "/api/account/current": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Get account information",
        "description": "Returns the user's non-sensitive account information.",
        "operationId": "api_current",
        "responses": {
          "200": {
            "description": "User's non-sensitive account information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentResponse"
                },
                "example": {
                  "budget_preference": "MediumBudget",
                  "disabilities": "knee replacement",
                  "email": "example@gmail.com",
                  "first_name": "First",
                  "food_allergies": "peanuts,vegetarian,pollen",
                  "last_name": "Last",
                  "risk_preference": "Adventurer"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "User has an invalid cookie/no cookie"
          },
          "405": {
            "description": "Method Not Allowed - Must be GET"
          },
          "408": {
            "description": "Request Timed Out"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "set-cookie": []
          }
        ]
      }
    },
    "/api/account/login": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Attempt user login",
        "description": "Attempts to login and return with a cookie.",
        "operationId": "api_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "example@gmail.com",
                "password": "Password_123"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login succeeded"
          },
          "400": {
            "description": "Bad Request"
          },
          "405": {
            "description": "Method Not Allowed - Must be POST"
          },
          "408": {
            "description": "Request Timed Out"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {},
          {
            "set-cookie": []
          }
        ]
      }
    },
    "/api/account/logout": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Logout by returning with expired cookie",
        "description": "Sets the HTTP-only cookie as expired, which deauthenticates the user.",
        "operationId": "api_logout",
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "User has an invalid cookie/no cookie"
          },
          "405": {
            "description": "Method Not Allowed - Must be GET"
          },
          "408": {
            "description": "Request Timed Out"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "set-cookie": []
          }
        ]
      }
    },
    "/api/account/signup": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Create a new account",
        "description": "Inserts account details into db (if email isn't already taken), and returns with a cookie.",
        "operationId": "api_signup",
        "requestBody": {
          "description": "Email must not already be taken. Password must be in plaintext.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              },
              "example": {
                "email": "example@gmail.com",
                "first_name": "First",
                "last_name": "Last",
                "password": "Password_123"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account successfully created"
          },
          "400": {
            "description": "Bad Request"
          },
          "405": {
            "description": "Method Not Allowed - Must be POST"
          },
          "408": {
            "description": "Request Timed Out"
          },
          "409": {
            "description": "Email already in use"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {},
          {
            "set-cookie": []
          }
        ]
      }
    },
    "/api/account/update": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Update information about the user",
        "description": "Update account info with provided data.",
        "operationId": "api_update",
        "requestBody": {
          "description": "Non-null fields will update that field. Null fields will not update that field.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequest"
              },
              "example": {
                "budget_preference": "LowBudget"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account info updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                },
                "example": {
                  "budget_preference": "LowBudget",
                  "disabilities": "knee replacement",
                  "email": "example@gmail.com",
                  "first_name": "First",
                  "food_allergies": "peanuts,vegetarian,pollen",
                  "last_name": "last",
                  "risk_preference": "Adventurer"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "User has an invalid cookie/no cookie"
          },
          "405": {
            "description": "Method Not Allowed - Must be POST"
          },
          "408": {
            "description": "Request Timed Out"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "set-cookie": []
          }
        ]
      }
    },
    "/api/account/validate": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Whether the user has a valid auth-token",
        "description": "Returns 200 if token is valid, or 401 if invalid or nonexistant.",
        "operationId": "api_validate",
        "responses": {
          "200": {
            "description": "User has a valid cookie"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "User has an invalid cookie/no cookie"
          },
          "405": {
            "description": "Method Not Allowed - Must be GET"
          },
          "408": {
            "description": "Request Timed Out"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "set-cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BudgetBucket": {
        "type": "string",
        "description": "Budget preference enum mapped to Postgres `budget_bucket`.\nUsed in account preferences and returned by account APIs.\n- Fields:\n  - Enum variants representing budget bands",
        "enum": [
          "VeryLowBudget",
          "LowBudget",
          "MediumBudget",
          "HighBudget",
          "LuxuryBudget"
        ]
      },
...
      "CurrentResponse": {
        "type": "object",
        "description": "API route response for GET `/api/account/current`.\n- Safe-to-return account profile for current user",
        "required": [
          "email",
          "first_name",
          "last_name",
          "food_allergies",
          "disabilities"
        ],
        "properties": {
          "budget_preference": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BudgetBucket",
                "description": "Optional budget enum"
              }
            ]
          },
          "disabilities": {
            "type": "string",
            "description": "Optional food and allergies preferences"
          },
          "email": {
            "type": "string",
            "description": "Email"
          },
          "first_name": {
            "type": "string",
            "description": "First name"
          },
          "food_allergies": {
            "type": "string",
            "description": "Optional food and allergies preferences"
          },
          "last_name": {
            "type": "string",
            "description": "Last name"
          },
          "risk_preference": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RiskTolerence",
                "description": "Optional risk enum"
              }
            ]
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "description": "Request payload for POST `/api/account/login`.",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email"
          },
          "password": {
            "type": "string",
            "description": "Plaintext password submitted by the user"
          }
        }
      },
...
      "RiskTolerence": {
        "type": "string",
        "description": "Risk tolerance enum mapped to Postgres `risk_tolerence`.\nUsed in account preferences and returned by account APIs.\n- Fields:\n  - Enum variants representing risk appetite",
        "enum": [
          "ChillVibes",
          "LightFun",
          "Adventurer",
          "RiskTaker"
        ]
      },
...
      "SignupRequest": {
        "type": "object",
        "description": "Request payload for POST `/api/account/signup`.\nValidated server-side before insert.",
        "required": [
          "email",
          "first_name",
          "last_name",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "description": "Plaintext password submitted by the user"
          }
        }
      },
...
      "UpdateRequest": {
        "type": "object",
        "description": "Request payload for POST `/api/account/update`.\n- Only `Some` fields are updated.",
        "properties": {
          "budget_preference": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BudgetBucket",
                "description": "Optional new budget enum"
              }
            ]
          },
          "disabilities": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional new disabilites\n* String is a comma-separated list of preferences"
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional new email"
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional new first name"
          },
          "food_allergies": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional new food and allergies preferences\n* String is a comma-separated list of preferences"
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional new last name"
          },
          "password": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional new plaintext password"
          },
          "risk_preference": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RiskTolerence",
                "description": "Optional new risk enum"
              }
            ]
          }
        }
      },
      "UpdateResponse": {
        "type": "object",
        "description": "API route response for POST `/api/account/update`.\n- Contains full updated account profile for convenience.",
        "required": [
          "email",
          "first_name",
          "last_name",
          "food_allergies",
          "disabilities"
        ],
        "properties": {
          "budget_preference": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BudgetBucket",
                "description": "Optional budget enum"
              }
            ]
          },
          "disabilities": {
            "type": "string",
            "description": "Optional disabilites\n* String is a comma-separated list of preferences"
          },
          "email": {
            "type": "string",
            "description": "Current email"
          },
          "first_name": {
            "type": "string",
            "description": "Current first name"
          },
          "food_allergies": {
            "type": "string",
            "description": "Optional food and allergies preferences\n* String is a comma-separated list of preferences"
          },
          "last_name": {
            "type": "string",
            "description": "Current last name"
          },
          "risk_preference": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RiskTolerence",
                "description": "Optional risk enum"
              }
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "set-cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth-token",
        "description": "An HTTP-only cookie which must encode a valid account id, expiration timestamp, and other information"
      }
    }
  },
  "security": [
    {},
    {
      "set-cookie": []
    }
  ],
...
}